\name{ugPlotServerStatus}
\alias{ugPlotServerStatus}
\title{Get background ugplot job server status}
\usage{
ugPlotServerStatus(name = "default")
}
\arguments{
\item{name}{Local server handle name.}
}
\value{
A list with server metadata and running state.
}
\description{
Get background ugplot job server status.
}
