% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ugPlot}
\alias{ugPlot}
\title{Open a window with the application}
\usage{
ugPlot(dataset = data.frame())
}
\arguments{
\item{dataset}{A data.frame to be automatically loaded. Leave empty to add later}
}
\value{
NULL when no error
}
\description{
Open a window with the application
}
\examples{
\donttest{
    ugPlot(dataset = mtcars)
}

}
