\name{ugPlotServerStop}
\alias{ugPlotServerStop}
\title{Stop a background ugplot job server}
\usage{
ugPlotServerStop(name = "default")
}
\arguments{
\item{name}{Local server handle name.}
}
\value{
Invisibly returns TRUE when a process was stopped.
}
\description{
Stop a background ugplot job server.
}
